+++ src/BigBlueButtonWebhooksController.php
@@ -12,9 +12,9 @@ class BigBlueButtonWebhooksController
     /**
      * Invoke controller method.
      *
-     * @param  \Illuminate\Http\Request $request
-     * @param  string|null $configKey
-     * @return \Illuminate\Http\Response
+     * @param  \Illuminate\Http\Request  $request
+     * @param  string|null  $configKey
+     * @return \Symfony\Component\HttpFoundation\Response
      */
     public function __invoke(Request $request, string $configKey = null)
     {
@@ -30,8 +30,6 @@ class BigBlueButtonWebhooksController
             'process_webhook_job' => config('bigbluebutton-webhooks.process_webhook_job'),
         ]);
 
-        (new WebhookProcessor($request, $webhookConfig))->process();
-
-        return response()->json(['message' => 'ok']);
+        return (new WebhookProcessor($request, $webhookConfig))->process();
     }
 }