+++ src/WebhookCall.php
@@ -9,17 +9,26 @@ use Spatie\WebhookClient\WebhookConfig;
 
 class WebhookCall extends Model
 {
+    /**
+     * @param  \Spatie\WebhookClient\WebhookConfig  $config
+     * @param  \Illuminate\Http\Request  $request
+     * @return \Spatie\WebhookClient\Models\WebhookCall
+     */
     public static function storeWebhook(WebhookConfig $config, Request $request): Model
     {
-        // payload is not proper JSON, rather is it split between three blocks
+        // bigblubutton payload is build in expectation of multiple events
         $payload = $request->input();
         // transform event
         if ($event = Arr::get($payload, 'event', null) and is_string($event)) {
             $payload['event'] = json_decode($event, true);
         }
-        // create
+        // take the headers form the top
+        $headers = self::headersToStore($config, $request);
+        // parse and return
         return self::create([
             'name' => $config->name,
+            'url' => $request->fullUrl(),
+            'headers' => $headers,
             'payload' => $payload,
         ]);
     }