+++ tests/BigBlueButtonWebhookCallTest.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace BinaryCats\BigBlueButtonWebhooks\Tests;
+namespace Tests;
 
 use BinaryCats\BigBlueButtonWebhooks\ProcessBigBlueButtonWebhookJob;
 use Illuminate\Support\Facades\Event;
@@ -14,6 +14,9 @@ class BigBlueButtonWebhookCallTest extends TestCase
     /** @var \Spatie\WebhookClient\Models\WebhookCall */
     public $webhookCall;
 
+    /**
+     * @return void
+     */
     public function setUp(): void
     {
         parent::setUp();
@@ -43,6 +46,7 @@ class BigBlueButtonWebhookCallTest extends TestCase
                 'timestamp' => '1591652302965',
                 'domain' => 'example.com',
             ],
+            'url' => '/webhooks/bigbluebutton',
         ]);
 
         $this->processBigblueButtonwEbhookJob = new ProcessBigBlueButtonWebhookJob($this->webhookCall);