+++ tests/IntegrationTest.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace BinaryCats\BigBlueButtonWebhooks\Tests;
+namespace Tests;
 
 use Illuminate\Support\Facades\Event;
 use Illuminate\Support\Facades\Route;
@@ -8,6 +8,11 @@ use Spatie\WebhookClient\Models\WebhookCall;
 
 class IntegrationTest extends TestCase
 {
+    /**
+     * @return void
+     *
+     * @throws \Exception
+     */
     public function setUp(): void
     {
         parent::setUp();
@@ -37,7 +42,6 @@ class IntegrationTest extends TestCase
 
         $webhookCall = WebhookCall::first();
 
-        $this->assertEquals('my.type', $webhookCall->payload['event'][0]['data']['id']);
         $this->assertEquals($payload, $webhookCall->payload);
         $this->assertNull($webhookCall->exception);