+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 
 All notable changes to `laravel-bigbluebutton-webhooks` will be documented in this file
 
+## 9.0.0 - 2022-01-26
+
+- Drop support for PHP 7
+- Upgrade spatie/laravel-webhook-client to version 3.0
+- Test Laravel 9 integration
+
 ## 1.0.0 - 2020-06-08
 
 - initial release